reworked async client connection

+++ jetty-ajp/src/main/java/org/eclipse/jetty/ajp/Ajp13Generator.java
@@ -487,7 +487,7 @@ public class Ajp13Generator extends AbstractGenerator
 
     /* ------------------------------------------------------------ */
     @Override
-    public long flushBuffer() throws IOException
+    public int flushBuffer() throws IOException
     {
         try
         {
